requirements.txt
# https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astro-sdk-python[snowflake]==1.7.0